+++ src/query.rs
@@ -16,8 +16,7 @@ use validator::Validate;
 /// ```rust
 /// use actix_web::{error, web, App, FromRequest, HttpResponse};
 /// use serde_derive::Deserialize;
-/// use actix_web_validator::{Query, QueryConfig};
-/// use validator::Validate;
+/// use actix_web_validator::{Query, QueryConfig, Validate};
 ///
 /// #[derive(Deserialize, Validate)]
 /// struct Info {
@@ -77,8 +76,7 @@ impl Default for QueryConfig {
 /// ```rust
 /// use actix_web::{web, App};
 /// use serde_derive::Deserialize;
-/// use actix_web_validator::Query;
-/// use validator::Validate;
+/// use actix_web_validator::{Query, Validate};
 ///
 /// #[derive(Debug, Deserialize)]
 /// pub enum ResponseType {
@@ -164,8 +162,7 @@ where
 /// ```rust
 /// use actix_web::{web, App};
 /// use serde_derive::Deserialize;
-/// use actix_web_validator::Query;
-/// use validator::Validate;
+/// use actix_web_validator::{Query, Validate};
 ///
 /// #[derive(Debug, Deserialize)]
 /// pub enum ResponseType {