+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#17](https://github.com/rambler-digital-solutions/actix-web-validator/issues/17): Add reexport of Validate trait from validator.
 
 ## [2.0.2] 2020-12-28
 ### Fixed